requirements_txt.txt
streamlit>=1.28.0
tabula-py>=2.8.0
pandas>=2.0.0
openpyxl>=3.1.0
java-installer>=1.0.0